third_party/blink/web_tests/external/wpt/workers/dedicated-worker-runtime-error-is-not-parse-error-expected.txt

This is a testharness.js-based test.
[FAIL] Script parse error dispatches plain Event at Worker
  assert_equals: Script parse error should dispatch Event, not ErrorEvent expected function "function Event() { [native code] }" but got function "function ErrorEvent() { [native code] }"
Harness: the test ran to completion.
